+++ pubgate/db/__init__.py
@@ -1,0 +1,16 @@
+from pubgate.db.user import User
+from pubgate.db.boxes import Inbox, Outbox
+
+
+
+# import flask_admin
+# from flask_admin.contrib.pymongo.view import ModelView
+async def register_admin(app):
+    # TODO patch flask admin
+    # admin = flask_admin.Admin(app, name='Pubgate Admin', template_mode='bootstrap3')
+    #
+    # # Add views
+    # admin.add_view(ModelView(User))
+    # admin.add_view(ModelView(Outbox))
+    # admin.add_view(ModelView(Inbox))
+    pass
