+++ pubgate/db/user.py
@@ -3,15 +3,15 @@ from simple_bcrypt import generate_password_hash
 # import flask_admin
 # from flask_admin.contrib.pymongo.view import ModelView
 from pubgate.utils.user import UserUtils
-from pubgate.db.models import Outbox, Inbox
+from pubgate.db.boxes import Outbox, Inbox
 from pubgate.db.managers import BaseManager
 
 
-def actor_clean(data):
+def actor_clean(data, striptags=False):
     return [item["activity"]["object"]["actor"] for item in data]
 
 
-def actor_clean_inbox(data):
+def actor_clean_inbox(data, striptags=False):
     return [item["activity"]["object"]["object"] for item in data]
 
 
@@ -68,7 +68,8 @@ class User(BaseModel, UserUtils, BaseManager):
     async def outbox_paged(self, request):
         filters = {
             "deleted": False,
-            "user_id": self.name
+            "user_id": self.name,
+            "activity.type": "Create"
         }
         return await self.get_ordered(request, Outbox, filters,
                                       self.activity_clean, self.outbox)
@@ -76,7 +77,8 @@ class User(BaseModel, UserUtils, BaseManager):
     async def inbox_paged(self, request):
         filters = {
             "deleted": False,
-            "users": {"$in": [self.name]}
+            "users": {"$in": [self.name]},
+            "activity.type": "Create"
         }
         return await self.get_ordered(request, Inbox, filters,
                                       self.activity_clean, self.inbox)
